#430 Move Program.Main into WebApplication.Run.
